requirements.txt
# Astro Runtime includes the following pre-installed providers packages: https://www.astronomer.io/docs/astro/runtime-image-architecture#provider-packag
astronomer-cosmos
apache-airflow-providers-snowflake==5.0.0
apache-airflow-providers-amazon